backend/requirement.txt
Flask==3.0.3
flask-restful==0.3.10
flask-cors==4.0.0
flask_httpauth==4.8.0
requests==2.32.3
influxdb==5.3.2
flasgger==0.9.7.1
gunicorn==23.0.0
mysql-connector-python==9.1.0
python-dotenv

prometheus-client==0.21.1
